dom/os/test/test_symlink.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>os symlink test</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1141021">os tests</a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test"></pre>

  <script>
  var jsFuns = SpecialPowers.Cu.getJSTestingFunctions();

  function runTest() {
    return new Promise((res, rej) => {
      var workerCode = `
        try {
          var path = navigator.os.TEMP_DIR + '/test_symlink';
          try {
            navigator.os.unlink(path);
          }
          catch (ex) {}

          navigator.os.symlink('whuuu', path);
          var rl = navigator.os.readlink(path);

          postMessage(JSON.stringify({
            rl: rl
          }));
        }
        catch (ex) {
          postMessage("symlink err: " + ex);
        }
        `;

      var workerBlob = new Blob([workerCode], { type: 'application/javascript' });
      var w = new Worker(URL.createObjectURL(workerBlob));

      w.onmessage = e => {
        is(e.data.indexOf('symlink err'), -1, 'ran successfully');
        if (e.data.indexOf('symlink err') !== -1) {
          dump(e.data + '\n');
          return res();
        }

        try {
          var data = JSON.parse(e.data);
        }
        catch(ex) {
          ok(false, 'is valid JSON (' + e.data + ')');
          return res();
        }

        is(data.rl, 'whuuu', 'rl');

        res();
      };
    });
  }

  SimpleTest.waitForExplicitFinish();

  SpecialPowers.pushPrefEnv({"set":[["dom.os.security.disabled", true]]},
    Promise.all([
      runTest()
    ]).then(() => {
      SimpleTest.finish();
    }));
  </script>
</body>
</html>
